Remove <FormattedDate>
Since Intl.DateTimeFormat is not very reliable on React Native Android,
user of FormattedDate will face bugs. The best way to prevent the bugs is
to not use it. Uses of FormattedDate is highly recommended to use a
proven date formatting library such as moment.

A notable bug is that Intl.DateTimeFormat always use UTC timezone to
format date as opposed to using device's timezone.